+++ packages/stream_core_flutter/lib/src/theme/semantics/stream_color_scheme.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:stream_core/stream_core.dart' show Standard;
 import 'package:theme_extensions_builder_annotation/theme_extensions_builder_annotation.dart';
 
-import '../../theme/primitives/stream_colors.dart';
 import '../primitives/internal/tokens/dark/stream_tokens.dart' as dark_tokens;
 import '../primitives/internal/tokens/light/stream_tokens.dart' as light_tokens;
+import '../primitives/stream_colors.dart';
 
 part 'stream_color_scheme.g.theme.dart';